requirements.txt
# This template is low-dependency. 
# By default there is no requirements added here.
# Add the requirements you need to this file.
pillow==10.2.0